Added concepTypes to columnsFile
Added that conceptTypes can be specified in the columnsFile. As a workaround because concepType detection can go wrong of the first record, as it is currently implemented in transmart.

Added that fully whitespace values are not inserted to the database.
Added that values longer than 255 chars (max lenght in I2B2 ) are truncated to 254 chars.